readaloud/requirements.txt
# Core TTS dependencies
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.35.0
accelerate>=0.20.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0

# System integration
pyperclip>=1.8.2
watchdog>=3.0.0
keyboard>=0.13.5
pynput>=1.7.6

# GUI (optional)
tkinter-tooltip>=2.0.0

# Utilities
click>=8.1.0
rich>=13.0.0
python-dotenv>=1.0.0

# StreamDeck integration
websocket-client>=1.6.0
psutil>=5.9.0

# Background service (optional)
python-daemon>=3.0.0

# Web interface
flask>=2.3.0
